Implement Update Operation For Successed Contracts But Had Some Exeption Error's

+++ Application/BackgroundWorker/Common/Events/SuccessedContractAdded.cs
@@ -2,6 +2,6 @@
 {
     public class SuccessedContractAdded
     {
-        public long SuccessedContractId { get; set; }
+        public int SuccessedContractId { get; set; }
     }
 }
